change_data using a function, that is executed by the Data agent
fix data_helper

extend data_test
